+++ kernel-rs/src/kalloc.rs
@@ -60,7 +60,7 @@ pub unsafe fn freerange(mut pa_start: *mut libc::c_void, mut pa_end: *mut libc::
     let mut p = ((pa_start as usize)
         .wrapping_add(PGSIZE as usize)
         .wrapping_sub(1usize)
-        & !(PGSIZE - 1 as i32) as usize) as *mut libc::c_char;
+        & !(PGSIZE - 1) as usize) as *mut libc::c_char;
     while p.offset(PGSIZE as isize) <= pa_end as *mut libc::c_char {
         kfree(p as *mut libc::c_void);
         p = p.offset(PGSIZE as isize)
