+++ kernel-rs/src/sleeplock.rs
@@ -30,7 +30,7 @@ impl Sleeplock {
             .initlock(b"sleep lock\x00" as *const u8 as *const libc::c_char as *mut libc::c_char);
         lk.name = name;
         lk.locked = 0 as u32;
-        lk.pid = 0 as i32;
+        lk.pid = 0;
 
         lk
     }
@@ -41,7 +41,7 @@ impl Sleeplock {
             .initlock(b"sleep lock\x00" as *const u8 as *const libc::c_char as *mut libc::c_char);
         (*self).name = name;
         (*self).locked = 0 as u32;
-        (*self).pid = 0 as i32;
+        (*self).pid = 0;
     }
 
     pub unsafe fn acquire(&mut self) {
@@ -57,7 +57,7 @@ impl Sleeplock {
     pub unsafe fn release(&mut self) {
         (*self).lk.acquire();
         (*self).locked = 0 as u32;
-        (*self).pid = 0 as i32;
+        (*self).pid = 0;
         wakeup(self as *mut Sleeplock as *mut libc::c_void);
         (*self).lk.release();
     }
