+++ kernel-rs/src/sysproc.rs
@@ -7,7 +7,7 @@ use crate::{
 
 pub unsafe fn sys_exit() -> usize {
     let mut n: i32 = 0;
-    if argint(0 as i32, &mut n) < 0 as i32 {
+    if argint(0, &mut n) < 0 {
         return usize::MAX;
     }
     exit(n);
@@ -25,7 +25,7 @@ pub unsafe fn sys_fork() -> usize {
 
 pub unsafe fn sys_wait() -> usize {
     let mut p: usize = 0;
-    if argaddr(0 as i32, &mut p) < 0 as i32 {
+    if argaddr(0, &mut p) < 0 {
         return usize::MAX;
     }
     wait(p) as usize
@@ -34,11 +34,11 @@ pub unsafe fn sys_wait() -> usize {
 pub unsafe fn sys_sbrk() -> usize {
     let mut addr: i32 = 0;
     let mut n: i32 = 0;
-    if argint(0 as i32, &mut n) < 0 as i32 {
+    if argint(0, &mut n) < 0 {
         return usize::MAX;
     }
     addr = (*myproc()).sz as i32;
-    if growproc(n) < 0 as i32 {
+    if growproc(n) < 0 {
         return usize::MAX;
     }
     addr as usize
@@ -46,7 +46,7 @@ pub unsafe fn sys_sbrk() -> usize {
 
 pub unsafe fn sys_sleep() -> usize {
     let mut n: i32 = 0;
-    if argint(0 as i32, &mut n) < 0 as i32 {
+    if argint(0, &mut n) < 0 {
         return usize::MAX;
     }
     tickslock.acquire();
@@ -64,7 +64,7 @@ pub unsafe fn sys_sleep() -> usize {
 
 pub unsafe fn sys_kill() -> usize {
     let mut pid: i32 = 0;
-    if argint(0 as i32, &mut pid) < 0 as i32 {
+    if argint(0, &mut pid) < 0 {
         return usize::MAX;
     }
     kill(pid) as usize
